Moved functions back to DWM1000 file
