--- src/DWM1000Test.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace DWM1000 {
-    /* Used for regulatory tx power testing */
-    void setTransmitPowerSpectrumTestMode(int32_t repeat_interval);
-}
